app/requirements.txt
streamlit==1.40.0
tensorflow==2.12.0
numpy==1.26.4
opencv-python==4.10.0.84
Pillow==10.3.0
h5py==3.10.0
